+++ plugins/modules/zhmc_cpc_capacity.py
@@ -451,7 +451,7 @@ from ansible.module_utils.basic import AnsibleModule  # noqa: E402
 from ..module_utils.common import log_init, open_session, close_session, \
     hmc_auth_parameter, Error, missing_required_lib, \
     common_fail_on_import_errors, parse_hmc_host, \
-    underscore_properties  # noqa: E402
+    underscore_properties, blanked_params  # noqa: E402
 
 try:
     import urllib3
@@ -835,9 +835,9 @@ def main():
 
     module.params['hmc_host'] = parse_hmc_host(module.params['hmc_host'])
 
-    _params = dict(module.params)
-    del _params['hmc_auth']
-    LOGGER.debug("Module entry: params: %r", _params)
+    if LOGGER.isEnabledFor(logging.DEBUG):
+        LOGGER.debug("Module entry: params: %r",
+                     blanked_params(module.params))
 
     try: